modules/video_stream/requirements.txt
# modules/video_stream/requirements.txt
paho-mqtt
websockets
numpy
opencv-python-headless
